Extract base abstract class for predicates
